Started working on IToolTip and its WinForms-based implementation
