+++ Source/OxyPlot/Utilities/IToolTip.cs
@@ -1,0 +1,14 @@
+﻿namespace OxyPlot
+{
+    using System;
+    using System.Collections.Generic;
+    using System.Text;
+
+    public interface IToolTip
+    {
+        void Show();
+        void Hide();
+        string Text { get; set; }
+        void Dispose();
+    }
+}
